test: complete residual access coverage and docs
